+++ tabulas-core/src/main/scala/de/tudresden/inf/lat/tabulas/extension/DefaultExtension.scala
@@ -24,11 +24,11 @@ class DefaultExtension extends Extension {
 
         val inputFileName= arguments(0)
         val outputFileName = arguments(1)
-        val tableMap: TableMap = new SimpleFormatParser(new FileReader(
+        val tableMap: TableMap = SimpleFormatParser(new FileReader(
           inputFileName)).parse()
         val output: BufferedWriter = new BufferedWriter(new FileWriter(
           outputFileName))
-        val renderer: SimpleFormatRenderer = new SimpleFormatRenderer(output)
+        val renderer: SimpleFormatRenderer = SimpleFormatRenderer(output)
         renderer.render(tableMap)
         true