+++ tabulas-ext/src/main/scala/de/tudresden/inf/lat/tabulas/ext/main/Main.scala
@@ -2,7 +2,7 @@ package de.tudresden.inf.lat.tabulas.ext.main
 
 import de.tudresden.inf.lat.tabulas.ext.parser.{CalendarParserExtension, CsvParserExtension, JsonParserExtension}
 import de.tudresden.inf.lat.tabulas.ext.renderer.{CsvExtension, HtmlExtension, JsonExtension, SqlExtension, WikitextExtension, YamlExtension}
-import de.tudresden.inf.lat.tabulas.extension.{DefaultExtension, NormalizationExtension}
+import de.tudresden.inf.lat.tabulas.extension.{DefaultExtension, NormalizationExtension, OldFormatExtension}
 import de.tudresden.inf.lat.tabulas.main.ConsoleStarter
 
 
@@ -17,6 +17,7 @@ object Main {
   def main(args: Array[String]): Unit = {
     val extensions = Seq(
       DefaultExtension(),
+      OldFormatExtension(),
       JsonParserExtension(),
       CsvParserExtension(),
       CalendarParserExtension(),