+++ tabulas-ext/src/main/scala/de/tudresden/inf/lat/tabulas/ext/parser/JsonParser.scala
@@ -24,7 +24,7 @@ class JsonParser(input: Reader) extends Parser {
   }
 
   def renderEntry(key: String, value: JsonValue): String = {
-    val prefix = key + ParserConstant.Space + ParserConstant.EqualsSign
+    val prefix = key + ParserConstant.Space + ParserConstant.EqualsFieldSign
     val middle = if (value.isNull) {
       ""
 
@@ -65,9 +65,9 @@ class JsonParser(input: Reader) extends Parser {
             .map(key => renderEntry(key, elements.get(key)))
             .mkString("")
           val newRecord = if (maybeMetadata.isDefined) {
-            ParserConstant.TypeSelectionToken + ParserConstant.Space + ParserConstant.EqualsSign + ParserConstant.Space + typeName
+            ParserConstant.TypeSelectionToken + ParserConstant.Space + ParserConstant.EqualsFieldSign + ParserConstant.Space + typeName
           } else {
-            ParserConstant.NewRecordToken + ParserConstant.Space + ParserConstant.EqualsSign + ParserConstant.Space
+            ParserConstant.NewRecordToken + ParserConstant.Space + ParserConstant.EqualsFieldSign + ParserConstant.Space
           }
           ParserConstant.NewLine + ParserConstant.NewLine + newRecord + ParserConstant.NewLine + recordStr
         }).mkString("") + ParserConstant.NewLine + ParserConstant.NewLine