+++ tabulas-ext/src/main/scala/de/tudresden/inf/lat/tabulas/ext/parser/JsonParserExtension.scala
@@ -27,7 +27,7 @@ class JsonParserExtension extends Extension {
           inputFileName)).parse()
         val output = new BufferedWriter(new FileWriter(
           outputFileName))
-        val renderer = new SimpleFormatRenderer(output)
+        val renderer = SimpleFormatRenderer(output)
         renderer.render(tableMap)
         true